Biologger v2 - Rising from the grave
Biologger has been horribly out of date for years. I've been meaning to clean
it up, but never got around to it. With the introduction of ChatGPT, I
decided to see if that could make something better. So most of this code
is now written by AI. Take that as you will. But it's working again and
even has a slightly updated look and feel.

Again, this is a quick and dirty project. It's not meant for production
use, and it's surely not an example of stellar design. It's also intended
for me, so it doesn't scale to multiple users. In other words, host it
yourself at your own risk and with no guarantees.